Hey Burton,
I have an example of a wig item. Just for fun. Apply the interface that you were speaking of as you like. Maybe make a branch and ill make a branch off of Develop - which i guess we should make now.

The WPF application is up. I did not create the Model-View-ViewModel folder structure, that will come in on a later commit.

In the tests Solution Folder, we needt o make a unit tests project, and play around with Square, possibly by passing around documents.

Feel free to text me or whatever, and i'll be doing work on this tomorrow.

Thanks Bud

- Endurance